pyproject.toml
[project]
name = "redesign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.60.0",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "mistralai>=1.9.3",
    "openai>=1.98.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "qdrant-client>=1.15.1",
    "tiktoken>=0.9.0",
    "unstructured>=0.18.11",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-experimental>=0.3.0",
    "pillow>=10.0.0",
    "neo4j>=5.28.2",
]
